loadelf: sign-extend ELF entry point into 64-bit PC
ms->pc was assigned from a uint32_t, zero-extending kernel addresses
(0x80xxxxxx) to positive 64-bit values instead of the correct negative
sign-extended form (0xffffffff80xxxxxx). Cast through int32_t so all
subsequent PC arithmetic in the interpreter — branch targets, J-type
upper-bit extraction, and pc+4 increments — naturally produces
sign-extended results matching the RTL.

+++ loadelf.cc
@@ -120,7 +120,7 @@ void load_elf(const char* fn, state_t *ms) {
   ph32 = reinterpret_cast<Elf32_Phdr*>(buf + e_phoff);
   e_shnum = bswap_(eh32->e_shnum);
   sh32 = reinterpret_cast<Elf32_Shdr*>(buf + bswap_(eh32->e_shoff));
-  ms->pc = lAddr;
+  ms->pc = (int64_t)(int32_t)lAddr;
 
   uint32_t loaddr = ~0U, hiaddr = 0;
   /* Find instruction segments and copy to
